+++ pages/settings/account/info/add-phone.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import AddPhone from '~/modules/Profile/Phone/AddPhone'
+import AddPhone from '~/modules/Profile/components/Phone/AddPhone'
 
 const AddPhonePage: FC = () => <AddPhone />
 export default AddPhonePage
