+++ pages/settings/account/info/email.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import Email from '~/modules/Profile/Email'
+import Email from '~/modules/Profile/components/Email'
 
 const EmailPage: FC = () => <Email />
 export default EmailPage
