+++ pages/settings/account/info/phone.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import Phone from '~/modules/Profile/Phone'
+import Phone from '~/modules/Profile/components/Phone'
 
 const PhonePage: FC = () => <Phone />
 export default PhonePage
